+++ docs/general/mcp/help.go
@@ -1,0 +1,7 @@
+package howai
+
+var Usage = []string{"mcp"}
+
+func GetDescription() string {
+	return "Start the JFrog MCP server and begin using it with your MCP client of your choice."
+}
